Update on DeckOfCards
This time around a GUI was built to interact with the DeckOfCards. So the progam starts with a deck of cards which we can choose to see a card following the order of the creation of the deck or we can choose to shuffle the deck then see a card, generally, not in the order that was created. Ignore the .jpg files those are just the images I used for the GUI.
Currently theres a bug where when the deck runs out of cards to show an error shows up on the console and new window opens. In the nearest future possible this will be corrected.